+++ __tests__/plots/static/index.ts
@@ -336,3 +336,11 @@ export { flareElementPointMoveAreaNormalizeY } from './flare-element-point-move-
 export { stackBarAnnotationStyleText } from './stack-bar-annotation-style-text';
 export { githubStarIntervalDarkThemeTransparent } from './github-star-interval-dark-theme-transparent';
 export { diamondIntervalGroupX } from './diamond-interval-group-x';
+export { mockIntervalSeriesOrder } from './mock-interval-series-order';
+export { mockPieSpider } from './mock-pie-spider';
+export { mockPieSpiderRight } from './mock-pie-spider-right';
+export { mockPieSpiderMany } from './mock-pie-spider-many';
+export { mockPieSpiderHide } from './mock-pie-spider-hide';
+export { mockPieSpiderExceed } from './mock-pie-spider-exceed';
+export { mockFacetPieLegend } from './mock-facet-pie-legend';
+export { weatherLineMultiAxesSync } from './weather-line-multi-axes-sync';